reddit-stream/requirements.txt
praw==7.4.0
kafka-python==2.0.2
APScheduler==3.7.0